third_party/blink/web_tests/wpt_internal/js-self-profiling/without-document-policy/disabled.https.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
  <script>
    promise_test(t => {
      const profiler = performance.profile({ sampleInterval: 10 });
      return promise_rejects_dom(t, 'NotAllowedError', profiler);
    }, 'profiling should throw without passing document policy');
  </script>
</body>
</html>
